Silencing a warning.

+++ src/benchmark.cpp
@@ -226,8 +226,8 @@ public:
         t2 = t1;
     }
     int elapsed() {
-        return static_cast<int> ((t2.tv_sec - t1.tv_sec) * 1000) + ((t2.tv_usec
-                - t1. tv_usec) / 1000);
+        return  (static_cast<int>(t2.tv_sec - t1.tv_sec) * 1000) + static_cast<int>(t2.tv_usec
+                - t1. tv_usec) / 1000;
     }
     int split() {
         gettimeofday(&t2, 0);
